Fix issue #75: Loss of CRT params on RSA keys
This fixes issue #75, which was caused by the fact that a call to
privateKey on RSAKey.Builder would not always use the CRT representation
of the private key. This change checks if a key is encoded using CRT,
and subsequently calls the proper function. Otherwise, the additional
parameters for the CRT encoding would get lost and the resulting PKCS8
encoding of the private keys would not match.